itertools

Module adaptors

Licensed under the Apache License, Version 2.0 https://www.apache.org/licenses/LICENSE-2.0 or the MIT license https://opensource.org/licenses/MIT, at your option. This file may not be copied, modified, or distributed except according to those terms.

Re-exports

pub use self::map::map_into;

pub use self::map::map_ok;

pub use self::map::MapInto;

pub use self::map::MapOk;

pub use self::coalesce::*;

pub use self::multi_product::*;

Modules

coalesce 🔒

map 🔒

multi_product 🔒

Macros

impl_tuple_combination 🔒

Structs

Batching

A “meta iterator adaptor”. Its closure receives a reference to the iterator and may pick off as many elements as it likes, to produce the next iterator element.

FilterMapOk

An iterator adapter to filter and apply a transformation on values within a nested Result::Ok.

FilterOk

An iterator adapter to filter values within a nested Result::Ok.

Interleave

An iterator adaptor that alternates elements from two iterators until both run out.

InterleaveShortest

An iterator adaptor that alternates elements from the two iterators until one of them runs out.

Positions

An iterator adapter to get the positions of each element that matches a predicate.

Product

An iterator adaptor that iterates over the cartesian product of the element sets of two iterators I and J.

PutBack

An iterator adaptor that allows putting back a single item to the front of the iterator.

TakeWhileRef

An iterator adaptor that borrows from a Clone-able iterator to only pick off elements while the predicate returns true.

Tuple1Combination

Tuple2Combination

Tuple3Combination

Tuple4Combination

Tuple5Combination

Tuple6Combination

Tuple7Combination

Tuple8Combination

Tuple9Combination

Tuple10Combination

Tuple11Combination

Tuple12Combination

TupleCombinations

An iterator to iterate through all combinations in a Clone-able iterator that produces tuples of a specific size.

Update

An iterator adapter to apply a mutating function to each element before yielding it.

WhileSome

An iterator adaptor that filters Option<A> iterator elements and produces A. Stops on the first None encountered.

Traits

HasCombination

Functions

batching

Create a new Batching iterator.

cartesian_product

Create a new cartesian product iterator

checked_binomial 🔒

filter_map_ok

Create a new FilterMapOk iterator.

filter_ok

Create a new FilterOk iterator.

interleave

Create an iterator that interleaves elements in i and j.

interleave_shortest

Create a new InterleaveShortest iterator.

positions

Create a new Positions iterator.

put_back

Create an iterator where you can put back a single item

take_while_ref

Create a new TakeWhileRef from a reference to cloneable iterator.

transpose_result 🔒

tuple_combinations

Create a new TupleCombinations from a cloneable iterator.

update

Create a new Update iterator.

while_some

Create a new WhileSome<I>.