Module itertools::adaptors::coalesce

Structs

CoalesceBy

DedupEq

DedupPred2CoalescePred

DedupPredWithCount2CoalescePred

NoCount

WithCount

Traits

CoalescePredicate

CountItem

DedupPredicate

Functions

coalesce

Create a new Coalesce.

dedup

Create a new Dedup.

dedup_by

Create a new DedupBy.

dedup_by_with_count

Create a new DedupByWithCount.

dedup_with_count

Create a new DedupWithCount.

Type Aliases

Coalesce

An iterator adaptor that may join together adjacent elements.

Dedup

An iterator adaptor that removes repeated duplicates.

DedupBy

An iterator adaptor that removes repeated duplicates, determining equality using a comparison function.

DedupByWithCount

An iterator adaptor that removes repeated duplicates, while keeping a count of how many repeated elements were present. This will determine equality using a comparison function.

DedupWithCount

An iterator adaptor that removes repeated duplicates, while keeping a count of how many repeated elements were present.