Type Alias itertools::adaptors::coalesce::DedupBy

pub type DedupBy<I, Pred> = CoalesceBy<I, DedupPred2CoalescePred<Pred>, NoCount>;

An iterator adaptor that removes repeated duplicates, determining equality using a comparison function.

See .dedup_by() or .dedup() for more information.

Aliased Type

struct DedupBy<I, Pred> {
    iter: I,
    last: Option<Option<<NoCount as CountItem<<I as Iterator>::Item>>::CItem>>,
    f: DedupPred2CoalescePred<Pred>,
}

Fields
iter: Ilast: Option<Option<<NoCount as CountItem<<I as Iterator>::Item>>::CItem>>

last is None while no item have been taken out of iter (at definition). Then last will be Some(Some(item)) until iter is exhausted, in which case last will be Some(None).

f: DedupPred2CoalescePred<Pred>