itertools::adaptors

Module map

Structs

MapSpecialCase

MapSpecialCaseFnInto

MapSpecialCaseFnOk

Traits

MapSpecialCaseFn

Functions

map_into

Create a new MapInto iterator.

map_ok

Create a new MapOk iterator.

Type Aliases

MapInto

An iterator adapter to apply Into conversion to each element.

MapOk

An iterator adapter to apply a transformation within a nested Result::Ok.