itertools

Module combinations

Structs

CombinationsGeneric

An iterator to iterate through all the k-length combinations in an iterator.

Traits

PoolIndex

A type holding indices of elements in a pool or buffer of items from an inner iterator and used to pick out different combinations in a generic way.

Functions

array_combinations

Create a new ArrayCombinations from a clonable iterator.

combinations

Create a new Combinations from a clonable iterator.

remaining_for ๐

For a given size n, return the count of remaining combinations or None if it would overflow.

Type Aliases

ArrayCombinations

Iterator for const generic combinations returned by .array_combinations()

Combinations

Iterator for Vec valued combinations returned by .combinations()