itertools

Module combinations_with_replacement

Structs

CombinationsWithReplacementGeneric

An iterator to iterate through all the n-length combinations in an iterator, with replacement.

Functions

array_combinations_with_replacement

Create a new `ArrayCombinationsWithReplacement`` from a cloneable iterator.

combinations_with_replacement

Create a new CombinationsWithReplacement from a cloneable iterator.

remaining_for 🔒

For a given size n, return the count of remaining combinations with replacement or None if it would overflow.

Type Aliases

ArrayCombinationsWithReplacement

Iterator for const generic combinations_with_replacement returned by .array_combinations_with_replacement()

CombinationsWithReplacement

Iterator for Box<[I]> valued combinations_with_replacement returned by .combinations_with_replacement()