itertools::combinations_with_replacement

Function remaining_for

fn remaining_for(n: usize, first: bool, indices: &[usize]) -> Option<usize>

For a given size n, return the count of remaining combinations with replacement or None if it would overflow.