Module itertools::cons_tuples_impl

Macros

impl_cons_iter

Structs

ConsTuples

An iterator that maps an iterator of tuples like ((A, B), C) to an iterator of (A, B, C).

Functions

cons_tuples

Create an iterator that maps for example iterators of ((A, B), C) to (A, B, C).