Enum itertools::Diff

pub enum Diff<I, J> where
    I: Iterator,
    J: Iterator,  {
    FirstMismatch(usize, PutBack<I>, PutBack<J>),
    Shorter(usize, PutBack<I>),
    Longer(usize, PutBack<J>),
}

A type returned by the diff_with function.

Diff represents the way in which the elements yielded by the iterator I differ to some iterator J.

Variants
FirstMismatch(usize, PutBack<I>, PutBack<J>)

The index of the first non-matching element along with both iterator's remaining elements starting with the first mis-match.

Shorter(usize, PutBack<I>)

The total number of elements that were in J along with the remaining elements of I.

Longer(usize, PutBack<J>)

The total number of elements that were in I along with the remaining elements of J.