Enum itertools::FoldWhile

pub enum FoldWhile<T> {
    Continue(T),
    Done(T),
}

An enum used for controlling the execution of .fold_while().

See .fold_while() for more information.

Variants
Continue(T)

Continue folding with this value

Done(T)

Fold is complete and will return this value

Methods

impl<T> FoldWhile<T>

pub fn into_inner(self) -> T

Return the value in the continue or done.

pub fn is_done(&self) -> bool

Return true if self is Done, false if it is Continue.