Enum itertools::Position

pub enum Position<T> {
    First(T),
    Middle(T),
    Last(T),
    Only(T),
}

A value yielded by WithPosition. Indicates the position of this element in the iterator results.

See .with_position() for more information.

Variants
First(T)

This is the first element.

Middle(T)

This is neither the first nor the last element.

Last(T)

This is the last element.

Only(T)

This is the only element.

Methods

impl<T> Position<T>

pub fn into_inner(self) -> T

Return the inner value.