Function itertools::process_results

pub fn process_results<I, F, T, E, R>(iterable: I, processor: F) -> Result<R, E>where
    I: IntoIterator<Item = Result<T, E>>,
    F: FnOnce(ProcessResults<'_, I::IntoIter, E>) -> R,

“Lift” a function of the values of an iterator so that it can process an iterator of Result values instead.

IntoIterator enabled version of Itertools::process_results.