Module itertools::free

Free functions that create iterator adaptors or call iterator methods.

The benefit of free functions is that they accept any IntoIterator as argument, so the resulting code may be easier to read.

Functions

all
Test whether the predicate holds for all elements in the iterable.

any
Test whether the predicate holds for any elements in the iterable.

chain
Create an iterator that first iterates i and then j.

enumerate
Iterate iterable with a running index.

fold
Perform a fold operation over the iterable.

interleave
Create an iterator that interleaves elements in i and j.

join
Combine all iterator elements into one String, seperated by sep.

kmerge
Create an iterator that merges elements of the contained iterators.

max
Return the maximum value of the iterable.

merge
Create an iterator that merges elements in i and j.

min
Return the minimum value of the iterable.

rev
Iterate iterable in reverse.

sorted
Collect all the iterable's elements into a sorted vector in ascending order.

zip
Iterate i and j in lock step.

zip_eq
Iterate i and j in lock step.