itertools

Module kmerge_impl

Structs

HeadTail 🔒

Head element and Tail iterator pair

KMergeBy

An iterator adaptor that merges an arbitrary number of base iterators according to an ordering function.

KMergeByLt

Traits

KMergePredicate

Functions

heapify 🔒

Make data a heap (min-heap w.r.t the sorting).

kmerge

Create an iterator that merges elements of the contained iterators using the ordering function.

kmerge_by

Create an iterator that merges elements of the contained iterators.

sift_down 🔒

Sift down element at index (heap is a min-heap wrt the ordering)

Type Aliases

KMerge

An iterator adaptor that merges an arbitrary number of base iterators in ascending order. If all base iterators are sorted (ascending), the result is sorted.