Module itertools::misc

A module of helper traits and iterators that are not intended to be used directly.

Structs

FlatTuples
A helper iterator that maps an iterator of tuples like ((A, B), C) to an iterator of (A, B, C).

Traits

AppendTuple
A helper trait for (x, y, z) ++ w => (x, y, z, w), used for implementing iproduct!.

GenericRange
GenericRange is implemented by Rust's built-in range types, produced by range syntax like a.., ..b or c..d.

IntoIteratorTuple
Apply IntoIterator on each element of a tuple.

MendSlice
A trait for items that can maybe be joined together.

Slice
A helper trait to let ZipSlices accept both &[T] and &mut [T].

ToFloat
Helper trait to convert usize to floating point type.