itertools

Module peeking_take_while

Macros

peeking_next_by_clone ๐

Structs

PeekingTakeWhile

An iterator adaptor that takes items while a closure returns true.

Traits

PeekingNext

An iterator that allows peeking at an element before deciding to accept it.

Functions

peeking_take_while

Create a PeekingTakeWhile