Struct itertools::GroupByLazy

pub struct GroupByLazy<K, I, F> where I: Iterator {
    // some fields omitted
}

GroupByLazy is the storage for the lazy grouping operation.

If the groups are consumed in their original order, or if each group is dropped without keeping it around, then GroupByLazy uses no allocations. It needs allocations only if several group iterators are alive at the same time.

This type implements IntoIterator (it is not an iterator itself), because the group iterators need to borrow from this value. It should be stored in a local variable or temporary and iterated.

See .group_by_lazy() for more information.