Struct itertools::ISlice

pub struct ISlice<I> {
    // some fields omitted
}

A sliced iterator.

Note: slicing an iterator is not constant time, and much less efficient than slicing for example a vector.

Example

use std::iter::repeat;

let mut it = repeat('a').slice(..3);
assert_eq!(it.count(), 3);

Methods

impl<I> ISlice<I> where I: Iterator

fn new<R: GenericRange>(iter: I, range: R) -> ISlice<I>