Struct itertools::PutBack

pub struct PutBack<A, I> {
    // some fields omitted
}

An iterator adaptor that allows putting back a single item to the front of the iterator.

Iterator element type is A

Methods

impl<A, I> PutBack<A, I>

fn new(it: I) -> PutBack<A, I>

Iterator element type is A

fn put_back(&mut self, x: A)

Put back a single value to the front of the iterator.

If a value is already in the put back slot, it is overwritten.