Struct itertools::Stride

pub struct Stride<'a, A> {
    // some fields omitted
}

Similar to the slice iterator, but with a certain number of steps (stride) skipped per iteration.

Does not support zero-sized A.

Iterator element type is &'a A

Methods

impl<'a, A> Stride<'a, A>

unsafe fn from_ptr_len(begin: *const A, nelem: uint, stride: int) -> Stride<'a, A>

Create Stride iterator from a raw pointer.

impl<'a, A> Stride<'a, A>

fn from_slice(xs: &'a [A], step: int) -> Stride<'a, A>

Create Stride iterator from a slice and the element step count.

If step is negative, start from the back.

Example

use itertools::Stride;

let xs = [0i, 1, 2, 3, 4, 5];

let mut front = Stride::from_slice(xs.as_slice(), 2);
assert_eq!(front[0], 0);
assert_eq!(front[1], 2);

let mut back = Stride::from_slice(xs.as_slice(), -2);
assert_eq!(back[0], 5);
assert_eq!(back[1], 3);

fn from_stride(it: Stride<'a, A>, step: int) -> Stride<'a, A>

Create Stride iterator from an existing Stride iterator

fn swap_ends(&mut self)

Swap the begin and end and reverse the stride, in effect reversing the iterator.