Struct itertools::structs::GroupBy

pub struct GroupBy<K, I, F> where
    I: Iterator,  { /* fields omitted */ }

GroupBy is the storage for the lazy grouping operation.

If the groups are consumed in their original order, or if each group is dropped without keeping it around, then GroupBy uses no allocations. It needs allocations only if several group iterators are alive at the same time.

This type implements IntoIterator (it is not an iterator itself), because the group iterators need to borrow from this value. It should be stored in a local variable or temporary and iterated.

See .group_by() for more information.