Struct itertools::structs::MapResults

#[must_use = "iterator adaptors are lazy and do nothing unless consumed"]

pub struct MapResults<I, F> { /* fields omitted */ }

An iterator adapter to apply a transformation within a nested Result.

See .map_results() for more information.