Struct itertools::structs::ProcessResults

#[must_use = "iterator adaptors are lazy and do nothing unless consumed"]
pub struct ProcessResults<'a, I, E: 'a> { /* fields omitted */ }

An iterator that produces only the T values as long as the inner iterator produces Ok(T).

Used by process_results, see its docs for more information.