Type Alias itertools::structs::Coalesce

pub type Coalesce<I, F> = CoalesceBy<I, F, NoCount>;

An iterator adaptor that may join together adjacent elements.

See .coalesce() for more information.

Aliased Type

struct Coalesce<I, F> {
    iter: I,
    last: Option<Option<<NoCount as CountItem<<I as Iterator>::Item>>::CItem>>,
    f: F,
}

Fields
iter: Ilast: Option<Option<<NoCount as CountItem<<I as Iterator>::Item>>::CItem>>

last is None while no item have been taken out of iter (at definition). Then last will be Some(Some(item)) until iter is exhausted, in which case last will be Some(None).

f: F