Trait itertools::PeekingNext

pub trait PeekingNext: Iterator {
    // Required method
    fn peeking_next<F>(&mut self, accept: F) -> Option<Self::Item>
       where F: FnOnce(&Self::Item) -> bool;
}

An iterator that allows peeking at an element before deciding to accept it.

See .peeking_take_while() for more information.

This is implemented by peeking adaptors like peekable and put back, but also by a few iterators that can be peeked natively, like the slice’s by reference iterator (std::slice::Iter).

Required Methods

fn peeking_next<F>(&mut self, accept: F) -> Option<Self::Item>where F: FnOnce(&Self::Item) -> bool,

Pass a reference to the next iterator element to the closure accept; if accept returns true, return it as the next element, else None.