itertools
Module traits

Traits helpful for using certain Itertools methods in generic contexts.

Traits

HomogeneousTuple

Implemented for homogeneous tuples of size up to 12.

IteratorIndex

Used by Itertools::get to know which iterator to turn different ranges into.