itertools

Module tuple_impl

Some iterator that produces tuples

Macros

impl_tuple_collect 🔒

rev_for_each_ident 🔒

Structs

CircularTupleWindows

An iterator over all windows, wrapping back to the first elements when the window would otherwise exceed the length of the iterator, producing tuples of a specific size.

TupleBuffer

An iterator over a incomplete tuple.

TupleWindows

An iterator over all contiguous windows that produces tuples of a specific size.

Tuples

An iterator that groups the items in tuples of a specific size.

Traits

HomogeneousTuple

Implemented for homogeneous tuples of size up to 12.

TupleCollect

Functions

add_then_div 🔒

(n + a) / d avoiding overflow when possible, returns None if it overflows.

circular_tuple_windows

tuple_windows

Create a new tuple windows iterator.

tuples

Create a new tuples iterator.