Module itertools::with_position

Structs

WithPosition

An iterator adaptor that wraps each element in an Position.

Enums

Position

A value yielded by WithPosition. Indicates the position of this element in the iterator results.

Functions

with_position

Create a new WithPosition iterator.