Crate itertools

Itertools — extra iterator adaptors, functions and macros

To use the macros in this crate, use the phase(plugin) attribute:

#![feature(phase)]
#[phase(plugin, link)] extern crate itertools;

To use the iterator methods in this crate, import the Itertools trait:

use itertools::Itertools;

You can shorten the crate name with something like:

use itertools as it;

License

Dual-licensed to be compatible with the Rust project.

Licensed under the Apache License, Version 2.0 http://www.apache.org/licenses/LICENSE-2.0 or the MIT license http://opensource.org/licenses/MIT, at your option. This file may not be copied, modified, or distributed except according to those terms.

Macros

icompr!
icompr as in “iterator comprehension” allows creating a mapped iterator with simple syntax, similar to set builder notation, and directly inspired by Python. Supports an optional filter clause.

Syntax:

icompr!(<expression> for <pattern> in <iterator>)

iproduct!
Create an iterator over the “cartesian product” of iterators.

izip!
Create an iterator running multiple iterators in lockstep.

Structs

Batching
An advanced iterator adaptor. The closure recives a reference to the iterator and may pick off as many elements as it likes, to produce the next iterator element.

Dedup
Remove duplicates from sections of consecutive identical elements. If the iterator is sorted, all elements will be unique.

FlatTuples
A helper iterator that maps an iterator of tuples like ((A, B), C) to an iterator of (A, B, C).

FnMap
Clonable iterator adaptor to map elementwise from Iterator<A> to Iterator<B>

GroupBy
Group iterator elements. Consecutive elements that map to the same key ("runs"), are returned as the iterator elements of GroupBy.

Interleave
Alternate elements from two iterators until both are run out

Intersperse
An iterator adaptor to insert a particular value between each element of the adapted iterator.

MapMut
Clonable iterator adaptor to map elementwise from Iterator<A> to Iterator<B>

Product
An iterator adaptor that iterates over the cartesian product of the element sets of two iterators I and J.

PutBack
An iterator adaptor that allows putting back a single item to the front of the iterator.

RcIter
A wrapper for Rc<RefCell<I>>, that implements the Iterator trait.

Stride
Stride is similar to the slice iterator, but with a certain number of steps (the stride) skipped per iteration.

StrideMut
StrideMut is like Stride, but with mutable elements.

Tee
One half of an iterator pair where both return the same elements.

Times
Iterator to repeat a simple number of times

ZipLongest
An iterator which iterates two other iterators simultaneously

Enums

EitherOrBoth
A value yielded by ZipLongest. Contains one or two values, depending on which of the input iterators are exhausted.

Traits

Itertools
Extra iterator methods for arbitrary iterators

Functions

linspace
Return an iterator with n elements, where the first element is a and the last element is b.

times
Return an iterator with n elements, for simple repetition a particular number of times. The iterator yields a counter.

write
Assign to each reference in to from from, stopping at the shortest of the two iterators.

Type Definitions

Linspace
An iterator of n evenly spaced floats.

Iterator element type is F.