Module itertools::adaptors::multi_product

Structs

MultiProduct
An iterator adaptor that iterates over the cartesian product of multiple iterators of type I.

MultiProductIter
Holds the state of a single iterator within a MultiProduct.

Enums

MultiProductIterState
Holds the current state during an iteration of a MultiProduct.

Functions

multi_cartesian_product
Create a new cartesian product iterator over an arbitrary number of iterators of the same type.