Enum itertools::EitherOrBoth

pub enum EitherOrBoth<A, B> {
    Both(A, B),
    Left(A),
    Right(B),
}

A value yielded by ZipLongest. Contains one or two values, depending on which of the input iterators are exhausted.

See .zip_longest() for more information.

Variants
Both(A, B)

Neither input iterator is exhausted yet, yielding two values.

Left(A)

The parameter iterator of .zip_longest() is exhausted, only yielding a value from the self iterator.

Right(B)

The self iterator of .zip_longest() is exhausted, only yielding a value from the parameter iterator.