Enum itertools::EitherOrBoth

pub enum EitherOrBoth<A, B> {
    Both(A, B),
    Left(A),
    Right(B),
}

A value yielded by ZipLongest. Contains one or two values, depending on which of the input iterators are exhausted.

Variants

Both

Neither input iterator is exhausted yet, yielding two values.

Left

The parameter iterator of .zip_longest() is exhausted, only yielding a value from the self iterator.

Right

The self iterator of .zip_longest() is exhausted, only yielding a value from the parameter iterator.