Function itertools::concat

pub fn concat<I>(iterable: I) -> I::Item where
    I: IntoIterator,
    I::Item: Extend<<<I as IntoIterator>::Item as IntoIterator>::Item> + IntoIterator + Default, 

Combine all an iterator’s elements into one element by using Extend.

IntoIterator-enabled version of .concat()

This combinator will extend the first item with each of the rest of the items of the iterator. If the iterator is empty, the default value of I::Item is returned.

use itertools::concat;
 
let input = vec![vec![1], vec![2, 3], vec![4, 5, 6]];
assert_eq!(concat(input), vec![1, 2, 3, 4, 5, 6]);