Module itertools::free

Free functions that create iterator adaptors or call iterator methods.

The benefit of free functions is that they accept any IntoIterator as argument, so the resulting code may be easier to read.

Functions

all

Test whether the predicate holds for all elements in the iterable.

any

Test whether the predicate holds for any elements in the iterable.

chain

Takes two iterables and creates a new iterator over both in sequence.

cloned

Create an iterator that clones each element from &T to T

enumerate

Iterate iterable with a running index.

fold

Perform a fold operation over the iterable.

interleave

Create an iterator that interleaves elements in i and j.

intersperse

Iterate iterable with a particular value inserted between each element.

intersperse_with

Iterate iterable with a particular value created by a function inserted between each element.

join

Combine all iterator elements into one String, separated by sep.

kmerge

Create an iterator that merges elements of the contained iterators using the ordering function.

max

Return the maximum value of the iterable.

merge

Create an iterator that merges elements in i and j.

merge_join_by

Return an iterator adaptor that merge-joins items from the two base iterators in ascending order.

min

Return the minimum value of the iterable.

multipeek

An iterator adaptor that allows the user to peek at multiple .next() values without advancing the base iterator.

peek_nth

A drop-in replacement for std::iter::Peekable which adds a peek_nth method allowing the user to peek at a value several iterations forward without advancing the base iterator.

put_back

Create an iterator where you can put back a single item

put_back_n

Create an iterator where you can put back multiple values to the front of the iteration.

rciter

Return an iterator inside a Rc<RefCell<_>> wrapper.

rev

Iterate iterable in reverse.

sorted

Sort all iterator elements into a new iterator in ascending order.

zipDeprecated

Converts the arguments to iterators and zips them.

zip_eq

Iterate i and j in lock step.