Module itertools::groupbylazy

Structs

Chunk

An iterator for the elements in a single chunk.

ChunkIndex

ChunkIndex acts like the grouping key function for IntoChunks

Chunks

An iterator that yields the Chunk iterators.

Group

An iterator for the elements in a single group.

GroupBy

GroupBy is the storage for the lazy grouping operation.

GroupInner

Groups

An iterator that yields the Group iterators.

IntoChunks

ChunkLazy is the storage for a lazy chunking operation.

Traits

KeyFunction

A trait to unify FnMut for GroupBy with the chunk key in IntoChunks

Functions

new

Create a new

new_chunks

Create a new