Module itertools::grouping_map

Structs

GroupingMap

GroupingMap is an intermediate struct for efficient group-and-fold operations. It groups elements by their key and at the same time fold each group using some aggregating operation.

GroupingMapFn

Functions

new

Creates a new GroupingMap from iter

new_map_for_grouping 🔒

Type Aliases

GroupingMapBy

GroupingMapBy is an intermediate struct for efficient group-and-fold operations.

MapForGrouping

A wrapper to allow for an easy into_grouping_map_by