itertools

Module merge_join

Structs

MergeBy

An iterator adaptor that merges the two base iterators in ascending order. If both base iterators are sorted (ascending), the result is sorted.

MergeFuncLR

MergeLte

Traits

FuncLR

OrderingOrBool

Functions

merge

Create an iterator that merges elements in i and j.

merge_by_new

Create a MergeBy iterator.

merge_join_by

Return an iterator adaptor that merge-joins items from the two base iterators in ascending order.

Type Aliases

Merge

An iterator adaptor that merges the two base iterators in ascending order. If both base iterators are sorted (ascending), the result is sorted.

MergeJoinBy

An iterator adaptor that merge-joins items from the two base iterators in ascending order.