Module itertools::sources

Iterators that are sources (produce elements from parameters, not from another iterator).

Structs

Iterate

An iterator that infinitely applies function to value and yields results.

RepeatCallDeprecated

See repeat_call for more information.

Unfold

See unfold for more information.

Functions

iterate

Creates a new iterator that infinitely applies function to value and yields results.

repeat_callDeprecated

An iterator source that produces elements indefinitely by calling a given closure.

unfold

Creates a new unfold source with the specified closure as the “iterator function” and an initial state to eventually pass to the closure