Struct itertools::Zip

pub struct Zip<T> { /* fields omitted */ }

An iterator that generalizes .zip() and allows running multiple iterators in lockstep.

The iterator Zip<(I, J, ..., M)> is formed from a tuple of iterators (or values that implement IntoIterator) and yields elements until any of the subiterators yields None.

The iterator element type is a tuple like like (A, B, ..., E) where A to E are the element types of the subiterator.

use itertools::Zip;

// Iterate over three sequences side-by-side
let mut xs = [0, 0, 0];
let ys = [69, 107, 101];

for (i, a, b) in Zip::new((0..100, &mut xs, &ys)) {
   *a = i ^ *b;
}

assert_eq!(xs, [69, 106, 103]);

Methods

impl<T> Zip<T> where
T: IntoIteratorTuple,
Zip<T::Output>: Iterator,

Important traits for Zip<(A,)>
impl<A> Iterator for Zip<(A,)> where
A: Iterator, type Item = (A::Item,);impl<A, B> Iterator for Zip<(A, B)> where
A: Iterator,
B: Iterator, type Item = (A::Item, B::Item);impl<A, B, C> Iterator for Zip<(A, B, C)> where
A: Iterator,
B: Iterator,
C: Iterator, type Item = (A::Item, B::Item, C::Item);impl<A, B, C, D> Iterator for Zip<(A, B, C, D)> where
A: Iterator,
B: Iterator,
C: Iterator,
D: Iterator, type Item = (A::Item, B::Item, C::Item, D::Item);impl<A, B, C, D, E> Iterator for Zip<(A, B, C, D, E)> where
A: Iterator,
B: Iterator,
C: Iterator,
D: Iterator,
E: Iterator, type Item = (A::Item, B::Item, C::Item, D::Item, E::Item);impl<A, B, C, D, E, F> Iterator for Zip<(A, B, C, D, E, F)> where
A: Iterator,
B: Iterator,
C: Iterator,
D: Iterator,
E: Iterator,
F: Iterator, type Item = (A::Item, B::Item, C::Item, D::Item, E::Item, F::Item);impl<A, B, C, D, E, F, G> Iterator for Zip<(A, B, C, D, E, F, G)> where
A: Iterator,
B: Iterator,
C: Iterator,
D: Iterator,
E: Iterator,
F: Iterator,
G: Iterator, type Item = (A::Item, B::Item, C::Item, D::Item, E::Item, F::Item, G::Item);impl<A, B, C, D, E, F, G, H> Iterator for Zip<(A, B, C, D, E, F, G, H)> where
A: Iterator,
B: Iterator,
C: Iterator,
D: Iterator,
E: Iterator,
F: Iterator,
G: Iterator,
H: Iterator, type Item = (A::Item, B::Item, C::Item, D::Item, E::Item, F::Item, G::Item, H::Item);impl<A, B, C, D, E, F, G, H, I> Iterator for Zip<(A, B, C, D, E, F, G, H, I)> where
A: Iterator,
B: Iterator,
C: Iterator,
D: Iterator,
E: Iterator,
F: Iterator,
G: Iterator,
H: Iterator,
I: Iterator, type Item = (A::Item, B::Item, C::Item, D::Item, E::Item, F::Item, G::Item, H::Item, I::Item);

pub fn new(t: T) -> Zip<T::Output>

Create a new Zip from a tuple of iterators.