Trait itertools::Itertools

pub trait Itertools<A>: Iterator<A> {
    fn fn_map<B>(self, map: fn(A) -> B) -> FnMap<A, B, Self> { ... }
    fn interleave<J: Iterator<A>>(self, other: J) -> Interleave<Self, J> { ... }
    fn intersperse(self, element: A) -> Intersperse<A, Self> { ... }
    fn clones(self) -> Clones<Self> { ... }
    fn dropn(&mut self, n: uint) -> uint { ... }
    fn drain(&mut self) { ... }
    fn apply(&mut self, f: |A|) { ... }
}

Extra iterator methods for arbitrary iterators

Provided Methods

fn fn_map<B>(self, map: fn(A) -> B) -> FnMap<A, B, Self>

Like regular .map, but using a simple function pointer instead, so that the resulting FnMap iterator value can be cloned.

Iterator element type is B

fn interleave<J: Iterator<A>>(self, other: J) -> Interleave<Self, J>

Alternate elements from two iterators until both are run out

Iterator element type is A

fn intersperse(self, element: A) -> Intersperse<A, Self>

An iterator adaptor to insert a particular value between each element of the adapted iterator.

Iterator element type is A

fn clones(self) -> Clones<Self>

An iterator like .map(|elt| elt.clone())

fn dropn(&mut self, n: uint) -> uint

Consume n elements of the iterator eagerly

Return actual number of elements consumed, until done or reaching the end.

fn drain(&mut self)

Run the iterator, eagerly, to the end and consume all its elements.

Example

use itertools::Itertools;

let mut cnt = 0i;
"hi".chars().map(|c| cnt += 1).drain();

fn apply(&mut self, f: |A|)

Run the closure f eagerly on each element of the iterator.

Consumes the iterator until its end.