Trait itertools::Itertools

pub trait Itertools: Iterator + Sized {
    fn fn_map<B>(self, map: fn(<Self as Iterator>::Item) -> B) -> FnMap<<Self as Iterator>::Item, B, Self> { ... }
    fn map_unboxed<B, F: FnMut(<Self as Iterator>::Item) -> B>(self, map: F) -> MapMut<F, Self> { ... }
    fn interleave<J: Iterator>(self, other: J) -> Interleave<Self, J> { ... }
    fn intersperse(self, element: <Self as Iterator>::Item) -> Intersperse<<Self as Iterator>::Item, Self> { ... }
    fn zip_longest<U: Iterator>(self, other: U) -> ZipLongest<Self, U> { ... }
    fn dedup(self) -> Dedup<<Self as Iterator>::Item, Self> { ... }
    fn batching<B, F: FnMut(&mut Self) -> Option<B>>(self, f: F) -> Batching<Self, F> { ... }
    fn group_by<K, F: FnMut(&<Self as Iterator>::Item) -> K>(self, key: F) -> GroupBy<<Self as Iterator>::Item, K, Self, F> { ... }
    fn tee(self) -> (Tee<<Self as Iterator>::Item, Self>, Tee<<Self as Iterator>::Item, Self>) { ... }
    fn into_rc(self) -> RcIter<Self> { ... }
    fn dropn(&mut self, n: uint) -> uint { ... }
    fn drain(&mut self) { ... }
    fn apply(&mut self, f: |<Self as Iterator>::Item|) { ... }
    fn collect_vec(self) -> Vec<<Self as Iterator>::Item> { ... }
}

Extra iterator methods for arbitrary iterators

Provided Methods

fn fn_map<B>(self, map: fn(<Self as Iterator>::Item) -> B) -> FnMap<<Self as Iterator>::Item, B, Self>

Like regular .map, but using a simple function pointer instead, so that the resulting FnMap iterator value can be cloned.

Iterator element type is B

fn map_unboxed<B, F: FnMut(<Self as Iterator>::Item) -> B>(self, map: F) -> MapMut<F, Self>

Like regular .map, but using an unboxed closure instead.

Iterator element type is B

fn interleave<J: Iterator>(self, other: J) -> Interleave<Self, J>

Alternate elements from two iterators until both are run out

Iterator element type is Self::Item

fn intersperse(self, element: <Self as Iterator>::Item) -> Intersperse<<Self as Iterator>::Item, Self>

An iterator adaptor to insert a particular value between each element of the adapted iterator.

Iterator element type is Self::Item

fn zip_longest<U: Iterator>(self, other: U) -> ZipLongest<Self, U>

Creates an iterator which iterates over both this and the specified iterators simultaneously, yielding pairs of two optional elements. When both iterators return None, all further invocations of next() will return None.

Example

let a = [0i];
let b = [1i, 2i];
let mut it = a.iter().cloned().zip_longest(b.iter().cloned());
assert_eq!(it.next(), Some(Both(0, 1)));
assert_eq!(it.next(), Some(Right(2)));
assert_eq!(it.next(), None);

Iterator element type is EitherOrBoth<Self::Item, B>.

fn dedup(self) -> Dedup<<Self as Iterator>::Item, Self>

Remove duplicates from sections of consecutive identical elements. If the iterator is sorted, all elements will be unique.

Iterator element type is Self::Item.

fn batching<B, F: FnMut(&mut Self) -> Option<B>>(self, f: F) -> Batching<Self, F>

An advanced iterator adaptor. The closure recives a reference to the iterator and may pick off as many elements as it likes, to produce the next iterator element.

Iterator element type is B.

Example

let xs = [0i, 1, 2, 1, 3];

// An adaptor that gathers elements up in pairs
let mut pit = xs.iter().cloned().batching(|mut it| {
           match it.next() {
               None => None,
               Some(x) => match it.next() {
                   None => None,
                   Some(y) => Some((x, y)),
               }
           }
       });
assert_eq!(pit.next(), Some((0, 1)));
assert_eq!(pit.next(), Some((2, 1)));
assert_eq!(pit.next(), None);

fn group_by<K, F: FnMut(&<Self as Iterator>::Item) -> K>(self, key: F) -> GroupBy<<Self as Iterator>::Item, K, Self, F>

Group iterator elements. Consecutive elements that map to the same key ("runs"), are returned as the iterator elements of GroupBy.

Iterator element type is (K, Vec<Self::Item>)

fn tee(self) -> (Tee<<Self as Iterator>::Item, Self>, Tee<<Self as Iterator>::Item, Self>)

Split into an iterator pair that both yield all elements from the original iterator.

The iterator element Self::Item must be clonable.

Iterator element type is Self::Item.

Example

let xs = vec![0i, 1, 2, 3];

let (mut t1, mut t2) = xs.into_iter().tee();
assert_eq!(t1.next(), Some(0));
assert_eq!(t1.next(), Some(1));
assert_eq!(t2.next(), Some(0));
assert_eq!(t1.next(), Some(2));
assert_eq!(t1.next(), Some(3));
assert_eq!(t1.next(), None);
assert_eq!(t2.next(), Some(1));

fn into_rc(self) -> RcIter<Self>

Return the iterator wrapped in a Rc<RefCell<_>> wrapper.

The returned RcIter can be cloned, and each clone will refer back to the same original iterator.

RcIter allows doing interesting things like using .zip on an iterator with itself, at the cost of runtime borrow checking. (If it is not obvious: this has a performance penalty.)

Example

let xs = [0i, 1, 1, 1, 2, 1, 3, 5, 6, 7];
let mut rit = xs.iter().cloned().into_rc();
let mut z = rit.clone().zip(rit.clone());
assert_eq!(z.next(), Some((0, 1)));
assert_eq!(z.next(), Some((1, 1)));
assert_eq!(z.next(), Some((2, 1)));
assert_eq!(rit.next(), Some(3));
assert_eq!(z.next(), Some((5, 6)));
assert_eq!(z.next(), None);

Panics in iterator methods if a borrow error is encountered, but it can only happen if the RcIter is reentered in for example .next(), i.e. if it somehow participates in an "iterator knot" where it is an adaptor of itself.

Iterator element type is Self::Item.

fn dropn(&mut self, n: uint) -> uint

Consume n elements of the iterator eagerly

Return actual number of elements consumed, until done or reaching the end.

fn drain(&mut self)

Run the iterator, eagerly, to the end and consume all its elements.

Example

use itertools::Itertools;

let mut cnt = 0i;
"hi".chars().map(|c| cnt += 1).drain();

fn apply(&mut self, f: |<Self as Iterator>::Item|)

Run the closure f eagerly on each element of the iterator.

Consumes the iterator until its end.

fn collect_vec(self) -> Vec<<Self as Iterator>::Item>

.collec_vec() is simply a type specialization of .collect(), for convenience.