Trait itertools::Itertools

pub trait Itertools: Iterator {
    fn fn_map<B>(self, map: fn(Self::Item) -> B) -> FnMap<B, Self> where Self: Sized { ... }
    fn interleave<J: Iterator<Item=Self::Item>>(self, other: J) -> Interleave<Self, J> where Self: Sized { ... }
    fn intersperse(self, element: Self::Item) -> Intersperse<Self> where Self: Sized, Self::Item: Clone { ... }
    fn zip_longest<U: Iterator>(self, other: U) -> ZipLongest<Self, U> where Self: Sized { ... }
    fn dedup(self) -> Dedup<Self> where Self: Sized { ... }
    fn batching<B, F: FnMut(&mut Self) -> Option<B>>(self, f: F) -> Batching<Self, F> where Self: Sized { ... }
    fn group_by<K, F: FnMut(&Self::Item) -> K>(self, key: F) -> GroupBy<K, Self, F> where Self: Sized { ... }
    fn tee(self) -> (Tee<Self>, Tee<Self>) where Self: Sized, Self::Item: Clone { ... }
    fn slice<R: GenericRange>(self, range: R) -> ISlice<Self> where Self: Sized { ... }
    fn into_rc(self) -> RcIter<Self> where Self: Sized { ... }
    fn step(self, n: usize) -> Step<Self> where Self: Sized { ... }
    fn merge<J>(self, other: J) -> Merge<Self, J> where Self: Sized, Self::Item: PartialOrd, J: Iterator<Item=Self::Item> { ... }
    fn cartesian_product<J>(self, other: J) -> Product<Self, J> where Self: Sized, Self::Item: Clone, J: Clone + Iterator { ... }
    fn enumerate_from<K>(self, start: K) -> EnumerateFrom<Self, K> where Self: Sized { ... }
    fn find_position<P>(&mut self, pred: P) -> Option<(usize, Self::Item)> where P: FnMut(&Self::Item) -> bool, Self: Sized { ... }
    fn dropn(&mut self, n: usize) -> usize { ... }
    fn dropping(self, n: usize) -> Self where Self: Sized { ... }
    fn drain(&mut self) where Self: Sized { ... }
    fn apply<F: FnMut(Self::Item)>(&mut self, f: F) where Self: Sized { ... }
    fn foreach<F: FnMut(Self::Item)>(&mut self, f: F) where Self: Sized { ... }
    fn collect_vec(self) -> Vec<Self::Item> where Self: Sized { ... }
    fn set_from<'a, A: 'a, J>(&mut self, from: J) -> usize where Self: Iterator<Item=&'a mut A>, J: Iterator<Item=A> { ... }
}

Extra iterator methods for arbitrary iterators

Provided Methods

fn fn_map<B>(self, map: fn(Self::Item) -> B) -> FnMap<B, Self> where Self: Sized

Like regular .map(), but using a simple function pointer instead, so that the resulting FnMap iterator value can be cloned.

Iterator element type is B.

fn interleave<J: Iterator<Item=Self::Item>>(self, other: J) -> Interleave<Self, J> where Self: Sized

Alternate elements from two iterators until both are run out

Iterator element type is Item.

fn intersperse(self, element: Self::Item) -> Intersperse<Self> where Self: Sized, Self::Item: Clone

An iterator adaptor to insert a particular value between each element of the adapted iterator.

Iterator element type is Item.

fn zip_longest<U: Iterator>(self, other: U) -> ZipLongest<Self, U> where Self: Sized

Creates an iterator which iterates over both this and the specified iterators simultaneously, yielding pairs of two optional elements. When both iterators return None, all further invocations of next() will return None.

Example

use itertools::EitherOrBoth::{Both, Right};
use itertools::Itertools;
let mut it = (0..1).zip_longest(1..3);
assert_eq!(it.next(), Some(Both(0, 1)));
assert_eq!(it.next(), Some(Right(2)));
assert_eq!(it.next(), None);

Iterator element type is EitherOrBoth<Item, B>.

fn dedup(self) -> Dedup<Self> where Self: Sized

Remove duplicates from sections of consecutive identical elements. If the iterator is sorted, all elements will be unique.

Iterator element type is Item.

fn batching<B, F: FnMut(&mut Self) -> Option<B>>(self, f: F) -> Batching<Self, F> where Self: Sized

A “meta iterator adaptor”. Its closure recives a reference to the iterator and may pick off as many elements as it likes, to produce the next iterator element.

Iterator element type is B.

Example

use itertools::Itertools;
// An adaptor that gathers elements up in pairs
let mut pit = (0..4).batching(|mut it| {
           match it.next() {
               None => None,
               Some(x) => match it.next() {
                   None => None,
                   Some(y) => Some((x, y)),
               }
           }
       });
assert_eq!(pit.next(), Some((0, 1)));
assert_eq!(pit.next(), Some((2, 3)));
assert_eq!(pit.next(), None);

fn group_by<K, F: FnMut(&Self::Item) -> K>(self, key: F) -> GroupBy<K, Self, F> where Self: Sized

Group iterator elements. Consecutive elements that map to the same key (“runs”), are returned as the iterator elements of GroupBy.

Iterator element type is (K, Vec<Item>)

fn tee(self) -> (Tee<Self>, Tee<Self>) where Self: Sized, Self::Item: Clone

Split into an iterator pair that both yield all elements from the original iterator.

Iterator element type is Item.

Example

use itertools::Itertools;
let xs = vec![0, 1, 2, 3];

let (mut t1, mut t2) = xs.into_iter().tee();
assert_eq!(t1.next(), Some(0));
assert_eq!(t1.next(), Some(1));
assert_eq!(t2.next(), Some(0));
assert_eq!(t1.next(), Some(2));
assert_eq!(t1.next(), Some(3));
assert_eq!(t1.next(), None);
assert_eq!(t2.next(), Some(1));

fn slice<R: GenericRange>(self, range: R) -> ISlice<Self> where Self: Sized

Return a sliced iterator.

Note: slicing an iterator is not constant time, and much less efficient than slicing for example a vector.

Iterator element type is Item.

Example

use std::iter::repeat;
use itertools::Itertools;

let mut it = repeat('a').slice(..3);
assert_eq!(it.count(), 3);

fn into_rc(self) -> RcIter<Self> where Self: Sized

Return an iterator inside a Rc<RefCell<_>> wrapper.

The returned RcIter can be cloned, and each clone will refer back to the same original iterator.

RcIter allows doing interesting things like using .zip() on an iterator with itself, at the cost of runtime borrow checking. (If it is not obvious: this has a performance penalty.)

Iterator element type is Item.

Example

use itertools::Itertools;

let mut rit = (0..9).into_rc();
let mut z = rit.clone().zip(rit.clone());
assert_eq!(z.next(), Some((0, 1)));
assert_eq!(z.next(), Some((2, 3)));
assert_eq!(z.next(), Some((4, 5)));
assert_eq!(rit.next(), Some(6));
assert_eq!(z.next(), Some((7, 8)));
assert_eq!(z.next(), None);

Panics in iterator methods if a borrow error is encountered, but it can only happen if the RcIter is reentered in for example .next(), i.e. if it somehow participates in an "iterator knot" where it is an adaptor of itself.

fn step(self, n: usize) -> Step<Self> where Self: Sized

Return an iterator adaptor that steps n elements in the base iterator for each iteration.

The iterator steps by yielding the next element from the base iterator, then skipping forward n - 1 elements.

Iterator element type is Item.

Panics if the step is 0.

Example

use itertools::Itertools;

let mut it = (0..8).step(3);
assert_eq!(it.next(), Some(0));
assert_eq!(it.next(), Some(3));
assert_eq!(it.next(), Some(6));
assert_eq!(it.next(), None);

fn merge<J>(self, other: J) -> Merge<Self, J> where Self: Sized, Self::Item: PartialOrd, J: Iterator<Item=Self::Item>

Return an iterator adaptor that merges the two base iterators in ascending order. If both base iterators are sorted (ascending), the result is sorted.

Iterator element type is Item.

Example

use itertools::Itertools;

let a = (0..10).step(2);
let b = (1..10).step(3);
let mut it = a.merge(b);
assert_eq!(it.next(), Some(0));
assert_eq!(it.next(), Some(1));
assert_eq!(it.next(), Some(2));
assert_eq!(it.next(), Some(4));
assert_eq!(it.next(), Some(4));
assert_eq!(it.next(), Some(6));

fn cartesian_product<J>(self, other: J) -> Product<Self, J> where Self: Sized, Self::Item: Clone, J: Clone + Iterator

Return an iterator adaptor that iterates over the cartesian product of the element sets of two iterators self and J.

Iterator element type is (Item, J::Item).

use itertools::Itertools;

let mut it = (0..2).cartesian_product("αβ".chars());
assert_eq!(it.next(), Some((0, 'α')));
assert_eq!(it.next(), Some((0, 'β')));
assert_eq!(it.next(), Some((1, 'α')));
assert_eq!(it.next(), Some((1, 'β')));
assert_eq!(it.next(), None);

fn enumerate_from<K>(self, start: K) -> EnumerateFrom<Self, K> where Self: Sized

Return an iterator adaptor that enumerates the iterator elements, starting from start and incrementing by one.

Iterator element type is (K, Self::Item).

use itertools::Itertools;

assert_eq!(
    "αβγ".chars().enumerate_from(-10i8).collect_vec(),
    [(-10, 'α'), (-9, 'β'), (-8, 'γ')]
);

fn find_position<P>(&mut self, pred: P) -> Option<(usize, Self::Item)> where P: FnMut(&Self::Item) -> bool, Self: Sized

Find the position and value of the first element satisfying a predicate.

fn dropn(&mut self, n: usize) -> usize

Consume the first n elements of the iterator eagerly.

Return actual number of elements consumed, until done or reaching the end.

fn dropping(self, n: usize) -> Self where Self: Sized

Consume the first n elements from the iterator eagerly, and return the same iterator again.

It works similarly to .skip(n) except it is eager and preserves the iterator type.

fn drain(&mut self) where Self: Sized

Deprecated: because of a name clash, use .count() or .foreach() instead as appropriate.

Run the iterator, eagerly, to the end and consume all its elements.

Example

use itertools::Itertools;

let mut cnt = 0us;
"hi".chars().map(|c| cnt += 1).drain();

fn apply<F: FnMut(Self::Item)>(&mut self, f: F) where Self: Sized

Run the closure f eagerly on each element of the iterator.

Consumes the iterator until its end.

Note: This method is deprecated, use .foreach() instead.

fn foreach<F: FnMut(Self::Item)>(&mut self, f: F) where Self: Sized

Run the closure f eagerly on each element of the iterator.

Consumes the iterator until its end.

fn collect_vec(self) -> Vec<Self::Item> where Self: Sized

.collect_vec() is simply a type specialization of .collect(), for convenience.

fn set_from<'a, A: 'a, J>(&mut self, from: J) -> usize where Self: Iterator<Item=&'a mut A>, J: Iterator<Item=A>

Assign to each reference in self from the from iterator, stopping at the shortest of the two iterators.

Return the number of elements written.

Example

use itertools::Itertools;

let mut xs = [0; 4];
xs.iter_mut().set_from(1..);
assert_eq!(xs, [1, 2, 3, 4]);