Module itertools::tuple_impl

Some iterator that produces tuples

Structs

TupleBuffer
An iterator over a incomplete tuple.

TupleWindows
An iterator over all contiguous windows that produces tuples of a specific size.

Tuples
An iterator that groups the items in tuples of a specific size.

Traits

TupleCollect

Functions

tuple_windows
Create a new tuple windows iterator.

tuples
Create a new tuples iterator.